subscriber/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-cloud-pubsub==2.21.1
google-cloud-bigquery==3.25.0
sqlalchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2